Рефакторинг структуры приложения: вынесены компоненты и константы
