+++ src/components/NotificationSnackbar.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Snackbar, Alert } from '@mui/material';
+
+const NotificationSnackbar = ({ open, message, severity, onClose }) => (
+  <Snackbar open={open} autoHideDuration={4000} onClose={onClose}>
+    <Alert onClose={onClose} severity={severity} sx={{ width: '100%' }}>
+      {message}
+    </Alert>
+  </Snackbar>
+);
+
+export default NotificationSnackbar;
